coworking/requirements.txt
Django>=4.0
djangorestframework>3.10.0
python-dateutil